Channel fields in CommonHandlers
Move channel fields into a base common handler.
