+++ eclair-core/src/main/scala/fr/acinq/eclair/channel/fsm/Channel.scala
@@ -163,7 +163,7 @@ object Channel {
 class Channel(val nodeParams: NodeParams, val wallet: OnChainChannelFunder, val remoteNodeId: PublicKey, val blockchain: typed.ActorRef[ZmqWatcher.Command], val relayer: ActorRef, val txPublisherFactory: Channel.TxPublisherFactory, val origin_opt: Option[ActorRef] = None)(implicit val ec: ExecutionContext = ExecutionContext.Implicits.global)
   extends FSM[ChannelState, ChannelData] with FSMDiagnosticActorLogging[ChannelState, ChannelData]
     with ChannelOpenSingleFunder
-    with UtilityHandlers
+    with CommonHandlers
     with FundingHandlers
     with ErrorHandlers {
