+++ eclair-core/src/main/scala/fr/acinq/eclair/channel/fsm/ChannelOpenSingleFunder.scala
@@ -17,10 +17,9 @@
 package fr.acinq.eclair.channel.fsm
 
 import akka.actor.typed.scaladsl.adapter.actorRefAdapter
-import akka.actor.{ActorRef, FSM, Status}
+import akka.actor.{FSM, Status}
 import akka.pattern.pipe
 import fr.acinq.bitcoin.{SatoshiLong, Script, ScriptFlags, Transaction}
-import fr.acinq.eclair.blockchain.OnChainChannelFunder
 import fr.acinq.eclair.blockchain.OnChainWallet.MakeFundingTxResponse
 import fr.acinq.eclair.blockchain.bitcoind.ZmqWatcher._
 import fr.acinq.eclair.channel.Helpers.{Funding, getRelayFees}
@@ -45,12 +44,7 @@ import scala.util.{Failure, Success, Try}
 /**
  * This trait contains the state machine for the single-funder channel funding flow.
  */
-trait ChannelOpenSingleFunder extends FSM[ChannelState, ChannelData] with FundingHandlers with UtilityHandlers with ErrorHandlers {
-
-  // @formatter:off
-  def wallet: OnChainChannelFunder
-  def origin_opt: Option[ActorRef]
-  // @formatter:on
+trait ChannelOpenSingleFunder extends FSM[ChannelState, ChannelData] with FundingHandlers with CommonHandlers with ErrorHandlers {
 
   when(WAIT_FOR_OPEN_CHANNEL)(handleExceptions {
     case Event(open: OpenChannel, d@DATA_WAIT_FOR_OPEN_CHANNEL(INPUT_INIT_FUNDEE(_, localParams, _, remoteInit, channelConfig, channelType))) =>
